fix: Merge suggestions and searchbox component files
