+++ src/components/search-box/search-box.stories.tsx
@@ -1,24 +1,12 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import { Configuration, ProductSearchOptions } from '@bloomreach/discovery-web-sdk';
+import { AutosuggestOptions, Configuration, ProductSearchOptions } from '@bloomreach/discovery-web-sdk';
 import { useContext } from 'react';
 import { LimitlessUIProvider } from '../../contexts/limitless-ui.provider';
 import { SearchContext } from '../../contexts/search.context';
 import { SearchBox } from './search-box';
 import { Theme } from '../theme';
-
-// Will be replaced with a 'results' component later
-const Results = () => {
-  const context = useContext(SearchContext);
-  return context?.searchResponse?.response?.docs?.map((result) => {
-    return (
-      <div key={result.pid}>
-        <h2>{result.title}</h2>
-        <p>{result.description}</p>
-      </div>
-    );
-  });
-};
+import { Suggestions } from '../suggestions';
 
 const meta: Meta<typeof SearchBox> = {
   title: 'SEARCH/SearchBox',
@@ -38,10 +26,18 @@ const meta: Meta<typeof SearchBox> = {
 export default meta;
 
 export type Story = StoryObj<typeof SearchBox>;
+export type SuggestionsStory = StoryObj<typeof Suggestions>;
 
 const configuration: Configuration = {
-  account_id: 6413,
-  domain_key: 'pacifichome',
+  account_id: 7634,
+  auth_key: 'zjlc0tsp2xu7l7ro',
+  domain_key: 'showcase_pacifichome',
+};
+
+const suggestOptions: Omit<AutosuggestOptions, 'q'> = {
+  _br_uid_2: 'test',
+  catalog_views: 'showcase_pacifichome',
+  url: 'https://example.com',
 };
 
 const searchOptions: Omit<ProductSearchOptions, 'q'> = {
@@ -51,6 +47,32 @@ const searchOptions: Omit<ProductSearchOptions, 'q'> = {
   start: 0,
   url: 'https://example.com',
   'facet.version': '3.0',
+};
+
+const searchBoxProps: SearchBoxProps = {
+  configuration,
+  searchOptions,
+  searchType: 'product',
+  labels: {
+    label: 'My basic label',
+    placeholder: 'Enter search here',
+    submit: 'Submit',
+    reset: 'Reset',
+  },
+};
+
+//
+// Will be replaced with a 'results' component later
+const Results = () => {
+  const context = useContext(SearchContext);
+  return context?.searchResponse?.response?.docs?.map((result) => {
+    return (
+      <div key={result.pid}>
+        <h2>{result.title}</h2>
+        <p>{result.description}</p>
+      </div>
+    );
+  });
 };
 
 export const Basic: Story = {
@@ -74,6 +96,24 @@ export const Basic: Story = {
       submit: 'Submit',
       reset: 'Reset',
     },
+  },
+};
+
+export const Autosuggest: Story = {
+  render: (args) => {
+    return (
+      <Theme>
+        <LimitlessUIProvider>
+          <SearchBox {...args}>
+            <Suggestions {...args} />
+          </SearchBox>
+        </LimitlessUIProvider>
+      </Theme>
+    );
+  },
+  args: {
+    configuration,
+    suggestOptions,
   },
 };
