--- src/components/suggestions/index.ts
@@ -1,3 +1,0 @@
-export { Suggestions } from './suggestions';
-export { useSuggestions } from '../../hooks/suggestions.hook';
-export * from './suggestions.types';
