+++ src/hooks/suggestions.hook.ts
@@ -1,9 +1,9 @@
 import { SuggestResponse } from '@bloomreach/discovery-web-sdk';
 import { useContext, useEffect, useMemo, useState } from 'react';
-import { debounce } from '../utils/debounce';
+import { SuggestionsProps } from '../components';
 import { AutoSuggestContext } from '../contexts/autosuggest.context';
+import { debounce } from '../utils/debounce';
 import { useAutoSuggest } from './autosuggest-api.hook';
-import { SuggestionsProps } from '../components/suggestions/suggestions.types';
 
 type UseSuggestions = {
   response: SuggestResponse | null;
